+++ settings.py
@@ -1,2 +1,3 @@
-ENTER_DELAY = 0
-WIN_R_DELAY = 0
+ENTER_DELAY = "0"
+WIN_R_DELAY = "0"
+file_extension = ".txt"