+++ changelog.md
@@ -1,9 +1,18 @@
+v0.4
+- Added the SHORT command, lets you make use of 2 key shortcuts (Win + R, Control + W, ect)
+- Added DELETE (will delete the file and end the program)
+- Command change CLEAR ( Clear will delete the current lines and let you type new ones from scratch )
+- Added extension to the setting file along with quick mode settings
+- Fixed some bugs
+- Updated README.md
+
 v0.3
 - Added READ command, it will print the whole file in the console
 - The file will now be saved each time a new command is entered. This will avoid data lost.
+
  -> This avoids lost of data if the program is closed unexpectedly but may use more resources
 - Change BREAK to STOP
 - Changed BLANK to DELETE
 - Fixed DELETE
-- Added setting folder in which the predetermined delays for ceratain commands can be changed
+- Added setting folder in which the predetermined delays for certain commands can be changed
 - Updated README.md